requirements.txt
opencv-python==4.8.1.78
flask==2.3.3
flask-restx==1.2.0
pyyaml==6.0.1
requests==2.31.0
google-generativeai>=0.7.0
pillow>=9.0.0
numpy==1.24.3